Make memory-card saves work on 64-bit hosts
The HSD memory-card library (hsd_3A94.c, hsd_3B27.c) stored pointers in
32-bit queue words and passed them through s32/int parameters, and the
second file wrote its 32-entry request queue into a 16-byte dummy array
because the console's contiguous work area had been split into separate
native globals. Answering Yes at the memory-card prompt therefore
corrupted memory instead of creating a save.

- Introduce CardWord (intptr_t on the host, s32 on console) for every
  command slot, request-queue entry and pointer-carrying parameter
  (icon/banner and header payload addresses, sub-file buffers,
  callbacks, filenames). Byte-offset accessors scale to word slots.
- Keep the header, 128 commands and 32 request entries in one shared
  work array declared in hsd_3A94.h.
- Clamp create-time block copies to the sub-file size, skip the bogus
  tenth descriptor slot, replace pointer-arithmetic block counting and
  the misaligned x170/x270 walk with plain indexing (native only).
- Widen lb_8001BE30/lb_8001BF04's icon and header pointers and store
  the icon manifest as bytes so host word order does not change it.
- Aurora GCI folder: missing files report NOFILE, delete and rename act
  on the host file, duplicate creates report EXIST, quieter lookups.
- native_hsd_card_test drives lbcardnew.c create, reload, read, rewrite
  and corruption detection through the card bridge onto an in-memory
  card under ASan/UBSan (Melee's 11-block, 90112-byte save layout).
- Device tooling: create/load input scripts, flip_save_trial.sh,
  check_gci.py, chunked push for the flaky tunnel. Docs now tell users
  to answer Yes and where the Dolphin-format .gci lives.

+++ src/melee/lb/lbcardnew.c
@@ -837,7 +837,7 @@ int lb_8001BA44(int chan, const char* filename, UNK_T status_out)
 }
 
 int lb_8001BB48(int chan, char* filename, void* file_entries, void* save_data,
-                const char* write_buf, int write_offset, int write_len,
+                const char* write_buf, intptr_t write_offset, intptr_t write_len,
                 UNK_T status_out)
 {
     int new_var;
@@ -932,7 +932,7 @@ int lb_8001BD34(int chan, const char* filename, UNK_T file_entries,
 #pragma dont_inline on
 #endif
 int lb_8001BE30(int chan, const char* filename, UNK_T file_entries,
-                const char* read_buf, int read_offset, int read_len,
+                const char* read_buf, intptr_t read_offset, intptr_t read_len,
                 UNK_T status_out, UNK_T callback)
 {
     struct CardTask* task;
@@ -965,7 +965,7 @@ int lb_8001BE30(int chan, const char* filename, UNK_T file_entries,
 #endif
 
 int lb_8001BF04(int chan, char* filename, void* file_entries,
-                const char* write_buf, int write_offset, int write_len,
+                const char* write_buf, intptr_t write_offset, intptr_t write_len,
                 UNK_T status_out)
 {
     struct CardTask* task;
