+++ src/melee/lb/lbcardnew.h
@@ -46,7 +46,7 @@ struct CardEntry {
                               UNK_T status_out);
 /* 01BB48 */ int lb_8001BB48(int chan, char* filename, void* file_entries,
                              void* save_data, const char* write_buf,
-                             int write_offset, int write_len,
+                             intptr_t write_offset, intptr_t write_len,
                              UNK_T status_out);
 /* 01BC18 */ int lb_8001BC18(int chan, char* filename, void** file_entries,
                              void* save_data, const char* write_buf,
@@ -56,11 +56,11 @@ struct CardEntry {
                                 UNK_T file_entries, UNK_T status_out);
 /* 01BE30 */ int lb_8001BE30(int chan, const char* filename,
                              UNK_T file_entries, const char* read_buf,
-                             int read_offset, int read_len, UNK_T status_out,
+                             intptr_t read_offset, intptr_t read_len, UNK_T status_out,
                              UNK_T callback);
 /* 01BF04 */ int lb_8001BF04(int chan, char* filename, void* file_entries,
-                             const char* write_buf, int write_offset,
-                             int write_len, UNK_T status_out);
+                             const char* write_buf, intptr_t write_offset,
+                             intptr_t write_len, UNK_T status_out);
 /* 01BFD8 */ int lb_8001BFD8(int chan,
                              lbCardNew_SnapshotEntry* snapshot_entries,
                              int* free_blocks, int* free_files);
